+++ src/it/unicam/cs/pa/lg/forza4/PlayerRandom.java
@@ -6,16 +6,16 @@ public class PlayerRandom extends Player
 {
 	private static final long serialVersionUID = -1123456076446536258L;
 
-	public PlayerRandom(final byte id)
+	public PlayerRandom(final int id)
 	{
 		super(id);
 	}
 	
 	@Override
-	public byte input()
+	public int input()
 	{
 		Random rng = new Random(System.currentTimeMillis());
-		return (byte) rng.nextInt(6);
+		return rng.nextInt(6);
 	}
